+++ datacube/scripts/product.py
@@ -151,8 +151,15 @@ def _write_tab(products):
 
 
 def _default_lister(products):
+    products = list(products)
+    if len(products) == 0:
+        return
+
+    max_w = max(len(p.name) for p in products)
+
     for prod in products:
-        echo(style(prod.name, fg='green') + '\t' + prod.definition.get('description', ''))
+        name = '{s:<{n}}'.format(s=prod.name, n=max_w)
+        echo(style(name, fg='green') + '  ' + prod.definition.get('description', ''))
 
 
 LIST_OUTPUT_WRITERS = {
@@ -181,15 +188,38 @@ def list_products(dc, output_format):
 @product_cli.command('show')
 @click.option('-f', 'output_format', help='Output format',
               type=click.Choice(['yaml', 'json']), default='yaml', show_default=True)
-@click.argument('product_name', nargs=1)
+@click.argument('product_name', nargs=-1)
 @ui.pass_datacube()
 def show_product(dc, product_name, output_format):
     """
     Show details about a product in the generic index.
     """
-    product = dc.index.products.get_by_name(product_name)
+
+    if len(product_name) == 0:
+        products = list(dc.index.products.get_all())
+    else:
+        products = []
+        for name in product_name:
+            p = dc.index.products.get_by_name(name)
+            if p is None:
+                echo('No such product: {!r}'.format(name), err=True)
+                sys.exit(1)
+            else:
+                products.append(p)
+
+    if len(products) == 0:
+        echo('No products', err=True)
+        sys.exit(1)
 
     if output_format == 'yaml':
-        yaml.dump(product.definition, sys.stdout, Dumper=SafeDatacubeDumper, default_flow_style=False, indent=4)
+        yaml.dump_all((p.definition for p in products),
+                      sys.stdout,
+                      Dumper=SafeDatacubeDumper,
+                      default_flow_style=False,
+                      indent=4)
     elif output_format == 'json':
+        if len(products) > 1:
+            echo('Can not output more than 1 product in json format', err=True)
+            sys.exit(1)
+        product, *_ = products
         click.echo_via_pager(json.dumps(product.definition, indent=4))